Update 20532C_LAB_AK_02.md (#113)
Removed the trailing '.' from Task 2, Step 7 e from the password field. Students commonly copy-paste the password, as the special characters ($) are a hassle for many international keyboards. This resulted in multiple students not being able to log in to a VM, as the password was incorrect when typed manually (but copy-pasted with the trailing dot succeeded).