Update 20532C_LAB_AK_04.md (#114)
Small tweaking to UI: Data + Storage is now Databases in the New-blade, so updated that. Also, the lab has a confusing bit at the start of Task 2: First, we click More Services, select SQL Databases but then we go to + New-blade. This confuses students, as the More Services > SQL Databases-blade also provides the Add -button near the + New-button. I did change the lab a bit to reflect the fact that one has to click the + New, not Add button.